processing-service/requirements.txt
fastapi==0.111.*
uvicorn==0.29.*
kafka-python==2.0.*
httpx==0.27.*
python-decouple==3.8
pytest==8.2.*
pytest-asyncio==0.23.*
pytest-cov==5.*